Step 06: Implement executing loaded ELF data

+++ boot/elf.h
@@ -1,6 +1,6 @@
 #ifndef H8_12STEPS_ELF_H
 #define H8_12STEPS_ELF_H
 
-int elf_load( uint8* buf );
+unsigned long elf_load( uint8* buf );
 
 #endif
